@@ -6,17 +6,20 @@
     "type": "project",
     "require": {
         "php": ">=5.5.9",
-        "laravel/framework": "5.1.*",
+        "laravel/framework": "5.2.*",
         "google/apiclient": "^1.1",
         "twbs/bootstrap": "^3.3",
         "jenssegers/mongodb": "^2.2",
-        "laravel/socialite": "^2.0"
+        "laravel/socialite": "^2.0",
+        "guzzlehttp/guzzle": "^6.2"
     },
     "require-dev": {
+        "symfony/dom-crawler": "~3.0",
         "fzaninotto/faker": "~1.4",
         "mockery/mockery": "0.9.*",
         "phpunit/phpunit": "~4.0",
-        "phpspec/phpspec": "~2.1"
+        "phpspec/phpspec": "~2.1",
+        "symfony/css-selector": "~3.0"
     },
     "autoload": {
         "classmap": [
