@@ -1,8 +1,9 @@
 <?php
 
 namespace App\Http\Middleware;
-
+use Auth;
 use Closure;
+use Session;
 use Illuminate\Contracts\Auth\Guard;
 
 class Authenticate
@@ -38,7 +39,7 @@ class Authenticate
             if ($request->ajax()) {
                 return response('Unauthorized.', 401);
             } else {
-                return redirect()->guest('auth/login');
+                return redirect()->guest('/');
             }
         }
 
