@@ -135,7 +135,7 @@ return [
     |
     */
 
-    'domain' => 'elastic-heart.dev',
+    'domain' => 'plexus.dev',
 
     /*
     |--------------------------------------------------------------------------
