@@ -14,9 +14,10 @@ class CreateUsers extends Migration
         Schema::create('users', function(Blueprint $table)
         {
             $table->increments('id');
-            $table->string('google');
-            $table->string('facebook');
+            $table->string('email');
+            $table->string('password');
             $table->string('name');
+            $table->string('avatar');
             $table->remembertoken();
             $table->timestamps();
         });
