@@ -12,17 +12,19 @@ class CreateEventsTable extends Migration
      */
     public function up()
     {
-            Schema::create('events', function(Blueprint $table) {
+        Schema::create('events', function(Blueprint $table) {
             $table->increments('id');
             $table->string('event_name');
             $table->string('event_des',1000);
             $table->string('soc_name');
             $table->timestamp('start_time');
             $table->timestamp('end_time');
+            $table->integer('duration')
             $table->integer('society_id')->length(10)->unsigned();
             $table->foreign('society_id')->references('id')->on('society')->onDelete('cascade');
             $table->tinyInteger('type');
             $table->boolean('approve');
+            $table->boolean('active');
         });
     }
 
@@ -34,5 +36,5 @@ class CreateEventsTable extends Migration
     public function down()
     {
         Schema::drop('events');
-   }
+    }
 }
