@@ -14,11 +14,12 @@ class CreateScoreTable extends Migration
     {
        Schema::create('scores', function(Blueprint $table) {
         $table->increments('id');
-        $table->integer('event_id')->length(10)->unsigned();
         $table->integer('user_id')->length(10)->unsigned();
-        $table->integer('score');
         $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade'); 
+        $table->integer('event_id')->length(10)->unsigned();
         $table->foreign('event_id')->references('id')->on('events')->onDelete('cascade');
+        $table->integer('score')->default(0);
+        $table->integer('level')->default(0);
     });
    }
 
