@@ -18,8 +18,8 @@
 
 <body>
     <div class = 'container'>
-        <a  href="{{ route('social_login', ['google']) }}"><img src="{{URL::asset('assets/img/login.png')}}"></a>
-        <a  href="{{route('social_login', ['facebook'])}}"><img src="{{URL::asset('assets/img/fblog.png')}}"></a>
+        <a  href="social/google"><img src="{{URL::asset('assets/img/login.png')}}"></a>
+        <a  href="social/facebook"><img src="{{URL::asset('assets/img/fblog.png')}}"></a>
      </div>
 </body>
 
